// src/schema/community/tests/editCommunity/editCommunity.generated.ts
/* eslint-disable */
/* @ts-nocheck */
/* prettier-ignore */
/* This file is automatically generated using `npm run graphql:types` */
import type * as Types from '../../../../generated/types';

import type { JsonObject } from "type-fest";
import gql from 'graphql-tag';
export type EditCommunityMutationVariables = Types.Exact<{
  input: Types.UpdateCommunityInput;
}>;


export type EditCommunityMutation = { __typename?: 'Mutation', editCommunity: { __typename?: 'Community', description: string | null, id: string, name: string | null, status: Types.CommnunityStatus } };


export const EditCommunity = gql`
    mutation EditCommunity($input: UpdateCommunityInput!) {
  editCommunity(input: $input) {
    description
    id
    name
    status
  }
}
    `;